Move computeWrappedTokenAddress and getNativeTokenInfo to the tokens context
