+++ src/contexts/tokens.context.tsx
@@ -14,15 +14,26 @@ import axios from "axios";
 
 import tokenIconDefaultUrl from "src/assets/icons/tokens/erc20-icon.svg";
 import { Erc20__factory } from "src/types/contracts/erc-20";
+import { Bridge__factory } from "src/types/contracts/bridge";
 import { getCustomTokens, cleanupCustomTokens } from "src/adapters/storage";
 import { useEnvContext } from "src/contexts/env.context";
 import { useErrorContext } from "src/contexts/error.context";
-import { useBridgeContext } from "src/contexts/bridge.context";
 import { useProvidersContext } from "src/contexts/providers.context";
 import { Chain, Env, Token } from "src/domain";
 import { getEthereumErc20Tokens } from "src/adapters/tokens";
 import { getEtherToken } from "src/constants";
 import * as ethereum from "src/adapters/ethereum";
+
+interface ComputeWrappedTokenAddressParams {
+  token: Token;
+  nativeChain: Chain;
+  otherChain: Chain;
+}
+
+interface GetNativeTokenInfoParams {
+  token: Token;
+  chain: Chain;
+}
 
 interface AddWrappedTokenParams {
   token: Token;
@@ -65,6 +76,11 @@ interface ApproveParams {
 
 interface TokensContext {
   tokens?: Token[];
+  computeWrappedTokenAddress: (params: ComputeWrappedTokenAddressParams) => Promise<string>;
+  getNativeTokenInfo: (params: GetNativeTokenInfoParams) => Promise<{
+    originNetwork: number;
+    originTokenAddress: string;
+  }>;
   addWrappedToken: (params: AddWrappedTokenParams) => Promise<Token>;
   getTokenFromAddress: (params: GetTokenFromAddressParams) => Promise<Token>;
   getToken: (params: GetTokenParams) => Promise<Token>;
@@ -76,6 +92,12 @@ interface TokensContext {
 const tokensContextNotReadyMsg = "The bridge context is not yet ready";
 
 const tokensContext = createContext<TokensContext>({
+  computeWrappedTokenAddress: () => {
+    return Promise.reject(tokensContextNotReadyMsg);
+  },
+  getNativeTokenInfo: () => {
+    return Promise.reject(tokensContextNotReadyMsg);
+  },
   addWrappedToken: () => {
     return Promise.reject(tokensContextNotReadyMsg);
   },
@@ -99,9 +121,49 @@ const tokensContext = createContext<TokensContext>({
 const TokensProvider: FC<PropsWithChildren> = (props) => {
   const env = useEnvContext();
   const { notifyError } = useErrorContext();
-  const { getNativeTokenInfo, computeWrappedTokenAddress } = useBridgeContext();
   const { connectedProvider, changeNetwork } = useProvidersContext();
   const [tokens, setTokens] = useState<Token[]>();
+
+  /**
+   * Provided a token, its native chain and any other chain, computes the address of the wrapped token on the other chain
+   */
+  const computeWrappedTokenAddress = useCallback(
+    async ({
+      token,
+      nativeChain,
+      otherChain,
+    }: ComputeWrappedTokenAddressParams): Promise<string> => {
+      const bridgeContract = Bridge__factory.connect(
+        otherChain.contractAddress,
+        otherChain.provider
+      );
+      return bridgeContract.precalculatedWrapperAddress(nativeChain.networkId, token.address);
+    },
+    []
+  );
+
+  /**
+   * Provided a token and a chain, when the token is wrapped, returns the native token's networkId and address and throws otherwise
+   */
+  const getNativeTokenInfo = useCallback(
+    ({
+      token,
+      chain,
+    }: GetNativeTokenInfoParams): Promise<{
+      originNetwork: number;
+      originTokenAddress: string;
+    }> => {
+      const bridgeContract = Bridge__factory.connect(chain.contractAddress, chain.provider);
+
+      return bridgeContract.wrappedTokenToTokenInfo(token.address).then((tokenInfo) => {
+        if (tokenInfo.originTokenAddress === ethersConstants.AddressZero) {
+          throw new Error(`Can not find a native token for ${token.name}`);
+        }
+        return tokenInfo;
+      });
+    },
+    []
+  );
 
   /**
    * Provided a token, if its property wrappedAddresses is missing, adds it and returns the new token
@@ -290,6 +352,8 @@ const TokensProvider: FC<PropsWithChildren> = (props) => {
   const value = useMemo(() => {
     return {
       tokens,
+      computeWrappedTokenAddress,
+      getNativeTokenInfo,
       getTokenFromAddress,
       getToken,
       getErc20TokenBalance,
@@ -299,6 +363,8 @@ const TokensProvider: FC<PropsWithChildren> = (props) => {
     };
   }, [
     tokens,
+    computeWrappedTokenAddress,
+    getNativeTokenInfo,
     getTokenFromAddress,
     getToken,
     getErc20TokenBalance,
